CHERI: Fix sigaction data structure for FFI
In C, the vast majority of the time the sigaction data structure
contains a union of function pointers as the first member of the struct.
In Rust, for the FFI definition of this same data structure we were
discarding a whole bunch of that information by representing that union
as a single usize.

Technically for targets where the size of usize is equal to the size of
a function pointer, this was ABI-compatible. However, for a CHERI target
we need to fix this mismatch in types for the ABI to work. included in
this fix is also a fix to expand that member out into a union of
function pointers so we also match the intended layout of the structure
from a programmer's point of view

The current state of this fix is slightly hacky because it is applied
arbitrarily to FreeBSD targets. It should be changed to either only
apply to CHERI targets, or this fix should be duplicated throughout.