CHERI: Define [u]intptr_t as pointers for Morello
In the CHERI C ABI, we are essentially dealing with capabilities when we
are passing [u]intptr_t around. The only difference is in the API/type
system, in that these can be treated as if they are integers.
Unfortunately, the ABI for an i128 as an underlying type for a
[u]intptr_t is not compatible with this, as it will get passed in two
integer registers instead of a capability register.

The compromise here is to treat [u]intptr_t entirely as if they are
pointers in Rust, since there is very minimal code that is doing
anything with these types other than just using them for FFI.