CHERI: Add epoll_event definition for Morello
With this definition we are taking into account the fact that the second
field in epoll_event actually is intended to represent a pointer. In the
Morello fork of Linux the type of this field has been changed to
uintptr_t. Currently this is a slight hack, ideally we'd also use
uintptr_t here but the pointer type is incapable of being sent between
threads safely so we use u128 instead...